Finishes new integration procedure.
Now, everything under solvers.co use the RK45 solver, even bisect. The actual computation of the new state has been isolated in the advanceStep routine, which ignores all the stuff about the new step computation. Furthermore, the detectin of the collision with the disk is done in the solver routine, still called SolverRK45 (TODO: change name, this is not just a RK45 solver). The code is now faster and more robust. At last, the 'resolution' variable has been deleted.